+++ dev/lckdclient.c
@@ -92,7 +92,7 @@ int main(int argc, char *argv[])
 				}
 
 				if (!strcmp(*args, "start") && len == 2) {
-					int port = 0;
+					uint16_t port = 0;
 					if(LOCKDOWN_E_SUCCESS == lockdownd_start_service(client, *(args + 1), &port)) {
 						printf("started service %s on port %i\n", *(args + 1), port);
 					}
